plat/kvm: Drop 'elf32-i386' output target linker flag
Remove the `elf32-i386` linker flag that forces the final Unikernel
binary output as a 32-bit kernel.

Since `qemu` currently boots us through `multiboot`, a boot protocol
that does not support 64-bit kernels, add a new support script
to be run on the final image. This `python` script will convert the
Unikernel's ELF64 headers to their ELF32 equivalents, thus fooling
the loader into thinking it's a 32-bit kernel. Since most of the
addresses contained into the fields of the headers are assumed to be of
at most 4 bytes, the script simply takes those values and directly places
them in their corresponding ELF32 header fields without truncating.

Since, at this time, the only supported boot protocol on x86 KVM is
`multiboot`, add a dependency on this script for x86 KVM builds.

This script will be used in the final image's build recipe: the image
will be renamed with `.ef64` suffix added so that in case the script
were to somehow fail along the way, the recipe will be triggered again
due to the missing target. The suffix is removed upon successful
completion

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>