plat/kvm: Drop `elf32-i386` output target linker flag
Remove the `elf32-i386` linker flag that forces the final Unikernel
binary output as a 32-bit kernel.

Since `qemu` currently boots us through `multiboot`, a boot protocol
that does not support 64-bit kernels, add a new support script
to be run on the final image. This `python` script will convert the
Unikernel's `ELF64` headers to their `ELF32` equivalents, thus fooling
the loader into thinking it's a 32-bit kernel. Since most of the
addresses contained into the fields of the headers are assumed to fit into
at most 4 bytes, the script simply takes those values and directly places
them in their corresponding `ELF32` header fields without truncating.

The actual `ELF32` header will be prepended to the final binary and the
equivalent `ELF32` Program/Section Headers will be appended to the end
of the binary so that we can keep the `Multiboot` header in the first
8192, as the specification requires.

The script will be called using a newly added `Makefile.rules` file meant
to contain definitions common to all platforms.

Since, at this time, the only supported boot protocol on x86 KVM QEMU is
`multiboot`, enable this script for x86 KVM QEMU Multiboot builds.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>