plat/common: Add macros for dynamic and relocation sections
Add `DYNAMIC_SECTIONS` macro to to represent `.dynamic` and its
related ELF sections. It does not matter to which segment we append
it to since we are not going to look at it at runtime, but for now
append it to the `data` segment. The size increase caused by the
addition of these necessary sections is insignificant (less than 512
bytes). Furthermore, for obvious reasons, append `.dynamic` to the
`dynamic` segment.

Add `RELOCATION_SECTIONS` macro for the `.rela.dyn` section.
Although different from the other `.rela.*` sections, we want
all of them to be contained into this one section, for easier
access, hence the `*(.rela.*)`.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>