plat/kvm/x86: Make the Unikernel position independent
Due to the high number of references to absolute symbol values in the
16-bit and 32-bit bootstrap code, a new macro, `ur_mov`, was introduced.
The new macro creates a new type of `_uk_reloc_` symbol, with an `imm`
suffix, to show that it is placed right after a `mov` instruction whose
immediate value must be relocated/patched. Thus, update `mkukreloc.py`
accordingly. One exception to this type of relocation is going to be
the 16-bit code, as there is no point in relocating it before SMP
initialization without even knowing the physical address in the
first 1MiB where the secondary cores are to start execution. So, add
an exception to these `start16` relocations in the script.

In order to make the 32-bit, no MMU, multiboot code relocatable,
make use of a locally defined `mb_r_quad` macro that simply forces
the placement of a `R_X86_64_64` relocation entry in `.rela.text` and
thus in `.rela.entry` (no need to generate a uk_reloc symbol just for
this) and write an x86 32-bit specific relocator that the multiboot
can call. Similarly to the 64-bit C code `do_uk_reloc` self relocator,
`do_uk_reloc32` is a macro that takes as argument a value equal or
different from 0 to specify whether a stack is supplied or not. If the
stack is not supplied, the macro can generate its own minimal scratch
stack. Furthermore, the base virtual address is expected in %esi:%edi
(lower 32 bits in %edi and higher 32 bits in %esi), as well as the
physical base address in the %edx register. Since we are in Protected
Mode, we can assume a 32-bit register is enough to hold this address.

Since, on this platform, we make use of static page tables, these were
also made relocatable through the `ur_pte` macro.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>