plat/common/x86: Implement fake PE header to be loadable through UEFI
Add a fake PE header to make UEFI think we are a valid UEFI application.
This relies on the `mkefi.py` script to fill in the required fields
after the final image is built.

Furthermore, this also implements the small x86 specific UEFI entry
point which will call the early self relocator, adjust the initially
unaligned stack and finally jump to the architecture generic EFI stub.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>