plat/common/arm: Implement fake PE header to be loadable through UEFI
Add a fake PE header to make UEFI think we are a valid UEFI application.
This relies on the `mkefi.py` script to fill in the required fields
after the final image is built.

Furthermore, this also implements the small ARM specific UEFI entry
point which will call the early self relocator and jump to the
architecture generic EFI stub.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>