plat/kvm/x86: Make the Unikernel position independent
In order to make the 32-bit, no MMU, `multiboot` code relocatable,
make use of a locally defined `mb_r_quad` macro that simply forces
the placement of a `R_X86_64_64` relocation entry in `.rela.text` and
thus in `.rela.entry` (no need to generate a ukreloc symbol just for
this) and use the x86 32-bit specific relocator as early as possible
to resolve all the necessary `ukreloc` entries.

As far as the rest of the x86 Unikernel goes, all `mov`'s with
absolute symbol references have been replaced accordingly with the
`ur_mov` instruction macro. One exception to this type of relocation
is going to be the 16-bit code, as there is no point in relocating
it before SMP initialization without even knowing the physical address
in the first 1MiB where the secondary cores are to start execution. So,
add an exception to these `start16` relocations in the script.

Since, on this platform, we make use of static page tables, these were
also made relocatable through the `ur_pte` macro.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>