plat/common: Add variable to force `.uk_reloc` in a loadable segment
Define a variable that is to be explicitly placed into the `.uk_reloc`
section. This way, the linker will be forced to map the section into a
loadable segment.

Add a new configuration option `MAX_UK_RELOC_SYMS` meant to
indicate the maximum amount of hand placed relocations generated by the
`ur_*` macro's. This helps avoid creating memory holes and spilling into
the `.bss` section during the update of `.uk_reloc` and stripping
of the `*dyn*` sections.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>