lib/ukreloc: Add inline function to apply a `ukreloc` entry
Implement a simple inline function to properly apply, depending on the
size of the relocation, a `ukreloc` entry.

`apply_ukreloc` may cause `UBSAN` to issue false positives, in
the case of the `x86` architecture which leads to the Unikernel crashing
since the serial console is not initialized yet, as it is mostly intended
to be called within the early self relocator. So make sure that
`UBSAN` does not touch `apply_ukreloc` because it makes sense, for
`x86`, to make unaligned accesses (especially in Kernel Space) in order
to resolve non-compliant relocations.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>