lib: Introduce `libukreloc`
In order to start supporting positional independence we need to be able
to self relocate. Thus, define a new custom relocation related structure
`struct ukreloc` that is meant to hold the offset in memory where the
relocation is to be applied, the offset of the value from the original
symbol value computed by the linker through the linker script's base
reference address, the size of the relocation and a flags field
respectively (for now we only have the `UKRELOC_FLAGS_PHYS_REL` flag
meant to indicate whether the relocation is to be applied based on a
physical address or not, which comes in handy in bootstrap code).

The final binary blob that the `mkukreloc.py` script builds is made of
a signature (UKRELOC_SIGNATURE), `struct ukreloc` entries appended
to each other and ends with a zeroed out `struct ukreloc` to act as a
sentinel. This binary blob will be placed in the binary through the
`.ukreloc` ELF section, updated through `objcopy`. Therefore, in order
for this forced section update to not mess up linker script exported
symbols, this section must be placed towards the end of the binary.

The binary blob is obtained by parsing all of the `.rela.dyn` entries,
converting them into `struct ukreloc` entries and by searching
through the debug image's symbols for `_ukreloc_` symbols created with
the help of the architecture independent helper assembly macro:
`ur_data`. This macro has been added to aid in replacing the references
to absolute symbol values with position independent equivalents that,
instead of referencing this very symbol, it places a placeholder value
(UKRELOC_PLACEHOLDER) and generates a symbol to be parsed by the script
to build a ukreloc entry. Furthermore, it increases the size of the
`.ukreloc` section by one entry through the `ur_sec_updt` macro. Note
that `ur_sec_updt` also inserts a dummy relocation entry in `.ukreloc`
in order to force the linker to not optimize away the symbols. A unique
`ukreloc` symbol will be generated through the `ur_sym` macro so that
an `ur_*` macro can be used on the same symbol more than once.
Consider the following usage example:
`ur_data quad, symbol, 8, _phys`
This will result in the `nm` tool, whose output the script will parse,
generating such entry:
`0000000000100106 T symbol_ukreloc_data8_phys`
For `mkukreloc.py` this means, with a linker script reference address
of 0x100000:
```
struct ukreloc {
        __u64 r_mem_off = 0x106;  // 0x100106 - 0x100000
        __u64 r_addr;
        __u32 r_sz = 8;  // symbol_ukreloc_data[8]_phys
        __u32 flags = UKRELOC_FLAGS_PHYS_REL;  // symbol_ukreloc_data8[_phys]
} __packed;
```
In order for actual value of `symbol` to be found, since its name is at
the beginning of the generated symbol, the script will now look for such
entry:
`0000000000100078 T symbol`  // from [symbol]_ukreloc_data8_phys
and thus, `0x78` becomes the value of `r_addr` and the
`struct ukreloc` is completed and appended to the binary blob with the
others.

`mkukreloc.py` will be invoked through `build_ukreloc` Makefile definition
and the "section.*lma.*adjusted to.*" type of `objcopy` warnings will be
ignored since they are intended and harmless. The script will also check
that the distance between `.ukreloc` and `.bss` is enough to not cause
an adjustment or spill. In case it is not enough, then the script will
raise an Exception.

All of this functionality shall be contained within the `libukreloc`
library.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>