include/uk/plat: Re-order memory region types by priority
Change the values of the memory region types so that they are ordered by
the priority through which one may decide to split or coalesce a memory
region. Being able to tell which memory region is of a higher importance
or more intangible, allows one, for example, to be able to tell when
comparing two overlapping regions of different types, whose overlapping
fragment to keep or discard.

For example, if one were to decide between two overlapping regions, one
of type `UKPLAT_MEMRT_RESERVED` and one of type `UKPLAT_MEMRT_FREE`,
the obvious choice would be to subtract the overlapping fragment from
the `UKPLAT_MEMRT_FREE` memory region. With this patch, the decision
will come through the comparison of types: the memory region is the
smaller type will have itself or its fragment removed.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>