support/scripts: Add `python3` script to patch fake PE header
This script allows patching of the architecture specific fake PE
headers. It only fills in the fields that UEFI firmware looks for
when validating and loading the image. Specifically, it does the
following:
- Write MS-DOS signature in the first bytes of the binary
- Write at the standard MS-DOS file offset `0x3c` the offset to the
beginning in file of the fake PE header
- Fill in the following fields of the Optional Header: SizeOfCode,
AddressOfEntryPoint, BaseOfCode, SizeOfImage
- Fill in the dummy PE sections, as PE, unlike ELF, is loaded by
sections:
	- dummy .reloc section pointing to itsel with all fields
	zeroed out except the VirtualAddress and PointerToRawData
	fields which point to the section itself, to fool UEFI into
	thinking this is a valid relocation.
	- .text section which contains the PT_LOAD segment with R/E
	permissions.
	- .rdata section which contains the PT_LOAD segment with RO
	permissions.
	- .data section which contains PT_LOAD segment with R/W
	permissions
	- .tdata section which contains PT_LOAD segment generated
	through `tls_load` segment, dedicated to `.tdata`
For these sections, only the following fields are required to be filed
in: VirtualSize, VirtualAddress, SizeOfRawData, PointerToRawData.

Thus, the script fills in the bare-minimum fields, according to EDKII,
the most complete and official UEFI implementation, that are required
by an UEFI application's PE header to be considered valid and loadable.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>