plat/common/x86: Add architecture specific post-EFI stub
Implement a stub that comes after the architecture generic EFI stub
calls the routine that exits `Boot Services`.

This is supposed to finish the architecture specific setupin order
to be able to have a Unikraft valid initial environment state before
`_ukplat_entry` is execited.

Thus, begin by disabling the `Interrupt Flag`, updating the root of
the page tables with our in-image static page tables, unmask the
legacy 8259 PIC, since, at this point, we do not have I/O APIC
support and, finally, jump to `lcpu_start64` with the entry function
as `_ukplat_entry` and a statically allocated page-sized stack.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>