plat/common: Generalize `bootmemory_palloc`
Since `bootmemory_palloc` can be used by any platform and
architecture and can come in handy when wanting to allocate a memory
region descriptor on the sport, that is not part of the original
memory map, move it to a more general location.

Furthermore, rename the function to `ukplat_memregion_alloc` to make
its usage-case more intuitive.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>