plat: Move `paging_init` to a platform common location
Since the `paging_init` function's implementation is fairly generic
and usable by other platforms/architectures, move it to a location
that is common to all.

Furthermore, make the addition of the memory region descriptor meant
to unmap the static boot page tables part of `paging_init`, since we
make use of static page tables everywhere, for now. Thus, as seen in
the static boot page tables, for x86 the first 4 GiB are statically
mapped and as far as ARM goes, memory up to the first 512 GiB is
mapped.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>