plat: Integrate `ukplat_memregion` API's into `ARM64` builds
Rely on the generic `ukplat_memregion` API's for memory region
management and drop `_libkvmplat_cfg` usage in favor of the
`bootinfo` getters.

Furthermore, now that `ukplat_memregion`'s are usable on `ARM64`,
proper paging support can be used instead of the overlycomplicated
memory re-adjustment that was previously done in `_init_dtb_mem`.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>