plat: Move paging initialization to a platform common location
Since the `paging_init` function of the x86 KVM subsystem is written
in a generic manner, it can be reused by other architectures or
platforms. Thus, move it to a platform generic location and rename
it with a more intuitive name to align it with the other function
declarations of `paging.c`.

Furthermore, since we may be placed at an address further than the
one assigned at link time, make sure to also unmap everything that
comes before us, except the first 1MiB, meant for the reserved
mappings. This way we are not losing sight of the memory regions
that come below the image itself.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>