plat/common/paging.c: Do not ignore `EEXIST` errors while mapping
f0ae44de725f ("plat/kvm/x86: Workaround re-mapping low-mem areas")
introduced this check to ignore `EEXIST` errors while mapping the
memory regions marked as `UKPLAT_MEMRF_MAP` due to the fact that,
before commit <generalize mem init commit sha here>, Unikraft was
ignoring all memory regions in the first 1 MiB and was manually
hardcoding the mappings in the static boot page tables. This was
done because the `multiboot` boot protocol would not report the
legacy mapped `BIOS ROM` area or the `VGA Framebuffer` in its
e820 map.

We no longer need this anymore and, if anything, it helps
us ensure that the setup of the boot environment's mappings is done
as we expect it to be.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>