lib/uksched: Introduce the per-thread auxiliary stack
Add a new field to `struct uk_thread` that can represent a secondary
stack that can be used as a backup stack.

This can become very useful in cases such as if wanting to defer
exception handling without creating another thread. For example,
we want to return from an exception into a function inside the
same thread to be able to do deferred I/O outside exception
context without contaminating the original stack that was present
before the trap. We can avoid polluting the original stack by using
this auxiliary stack instead.

Update `_uk_thread_struct_init` to also *internally* allocate the
auxiliary stack with a size given by the user while configuring
the unikernel: `CONFIG_LIBUKSCHED_AUXSP_SIZE`.

Therefore, since this allocation can fail, allow
`_uk_thread_struct_init` to return an error value and, subsequently,
update every caller of this function to be aware of this returned
value.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>