plat/kvm/x86: Early init `KERNEL_GS_BASE` to point to `struct lcpu`
Make an early initialization of the `KERNEL_GS_BASE` MSR to point to
current LCPU's `struct lcpu`. This will make it easy for us to access
`struct lcpu` fields without dirtying any registers or the stack
during fragile execution context where we may need to not modify
the context at all in order to preserve it by using `swapgs`.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>