arch: Introduce `struct ukarch_ulctx`
Define an architecture specific userland context that an architecture
can use to store things it may want to preserve or switch between
when entering/exiting Unikraft context during normal userland
execution (e.g. running in conjunction with binary syscalls).

For now, this is defined as a struct with two base fields for both
ARM64 and x86_64:
- `r`, that refers to the registers the syscall caller had before invoking
the syscall itself
- `flags`, that refer to various flags that may be used to change the
behaviour and/or interpretation of the userland context
(For now, only the `UKARCH_ULCTX_FLAGS_INSYSCALL` flag is defined and it
is meant to say that the current thread is in syscall context - this can
be used when rescheduling a thread that was put to sleep during a
syscall, e.g. `futex` and can help us differentiate between a thread
that whose current context is in that of a syscall and one which is
not)

Additionally, x86_64 has one more specfic field:
- `gs_base`, that refers to the value of the `gs_base` register of
the application before calling the system call

Lastly, allow each thread to have their own userland context,
`struct ukarch_ulctx`, which userland Unikraft can use
to properly switch contexts between the application and Unikraft
whenever a system call is issued. By having a per-thread binary
compatibility context we do not lose application threads states
between thread switches.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>