arch/x86: Add userland context `store`/`load` operations
Implement `store`/`load` functions to be used in conjunction with the
userland context.

The `store` operation assumes that it can only be called from within
Unikraft context (i.e. not directly by the app) and, as such, the
app's `gs_base` register value was preserved within
`X86_MSR_KERNEL_GS_BASE` following a `swapgs` on syscall entry.
This value will be thus stored in the current userland
context.

The `load` operation assumes the same thing and, therefore, it will
set `X86_MSR_GS_BASE` to the current `struct lcpu` and
`X86_MSR_KERNEL_GS_BASE` to the app's preserved `gs_base` register
within the userland context of the app's thread we
are currently switching to.

Notice that the `UKARCH_ULCTX_FLAGS_INSYSCALL` is used so that these
operations are ignored if we are not looking at a thread whose exection
context is taking place inside a legitimate system call context,
as there is no point to them outside system call context.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>