lib/syscall_shim/arch/x86_64: Add `ulctx` `switchon`/`switchoff`
Implement `switchoff`/`switchon` functions to be used in conjunction
with the userland context.

The `switchoff` operation assumes that it can only be called from within
Unikraft context (i.e. not directly by the app) and, as such, the
app's `gs_base` register value was preserved within
`X86_MSR_KERNEL_GS_BASE` following a `swapgs` on syscall entry.
This value will be thus stored in the current userland
context.

The `switchon` operation assumes the same thing and, therefore, it will
set `X86_MSR_GS_BASE` to the current `struct lcpu` and
`X86_MSR_KERNEL_GS_BASE` to the app's preserved `gs_base` register
within the userland context of the app's thread we are currently
switching to.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>