lib/uksched: Also `store`/`load` `ulctx` on thread switch
Just like the general purpose context or the extended context, we also
have to `store`/`load` the userland context whenever we
switch from one thread to another. This helps us maintain a proper state
in situations such as switching to a thread whose execution state was
within a system call.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>