arch/x86: Add `UKARCH_ULCTX_FLAGS_NEWCLONE` bitfield to ulctx flags
Add a bitfield to let us know if a userland context
appertains to a thread that has just been created through a clone()
system call and has yet to be scheduled to run.

If this bit is set, then `ukarch_ulctx_load` is actually executed
right before scheduling such a thread. Therefore, set the MSRs
accordingly: kernel gs_base to our `struct lcpu *` and regular
gs_base to the parent's. This needs to be done only once, thus
undo this flag afterwards and add an assertion for this flag in
`ukarch_ulctx_store` because it should not exist outside the first
`ukarch_ulctx_load`.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>