lib/posix-process: Copy full parent registers context on `clone`
Allow `clone` to make use of the userland context in order
to be able to give the child the same exact register context
as the parent, except, of course, for the obvious registers
(e.g. `%rax` equal to `0` on `x86_64` and `x0` equal to `0`
on `ARM64`).

Some runtimes, such as the Go one, may expect some registers
to be exact the same as those of the parent.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>