plat/common: Move `ukplat_lcpu_[id|current]` to arch specific files
Now that we have a per-cpu system register that holds the current
pointer to the `struct lcpu` element of the global `lcpus` array,
make use of that for each architecture independently and move
the implementations of `ukplat_lcpu_[id|current]` from the
architecture agnostic place to the more architecture specific ones.

NOTE: `lcpu_get_current` of `x86` still makes use of the now
optimized through `%gs` register by calling `ukplat_lcpu_idx`
which calls `ukplat_lcpu_id`. However, in the future, it
will be better if we make use of the `rdgsbase` instruction
and the like.

Co-authored-by: Michalis Pappas <michalis@unikraft.io>
Co-authored-by: Marco Schlumpp <marco@unikraft.io>
Signed-off-by: Michalis Pappas <michalis@unikraft.io>
Signed-off-by: Marco Schlumpp <marco@unikraft.io>
Signed-off-by: Sergiu Moga <sergiu@unikraft.io>