plat/kvm/arm: No IRQ's during traps, unless it is a syscall
Do not allow the occurrence of IRQ's during traps, especially early
on when we are saving registers, by disabling them entirely.

System calls, on the other hand, may be executed slowly and may put
the current thread to sleep waiting for an IRQ. Thus, give system
calls the chance to handle such cases by explicitly enabling IRQ's
if the current trap is a system call.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>