arch/x86/x86_64: Allow asm files access to register offsets macros
It may come in handy to be able to know the offsets of each register
inside the `struct __regs` structure during more fragile states of
execution, usually coded in assembly.

Therefore, move the guarding `#ifdef` below the register definitions
so that assembly files may be able to include this header and benefit
from the aforementioned macros.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>