lib/syscall_shim/arm: Define `struct uk_syscall_regs` saving prologue
Implement an ARM inline assembly macro `UK_SYSCALL_USR_PROLOGUE_DEFINE`
function named after `pname` that switches to the auxiliary stack,
aligns it to the desired `struct uk_syscall_regs` alignment and starts
pushing and storing the caller's context before passing it as argument
to a function whose name is defined by the second argument, `fname`.

This function shall be defined as a `__naked` function, meaning the
compiler provides us with not ABI compliant prologue/epilogue.
Althrough we supposedly don't touch the callee preserved registers in
this function, fully optimized images may inline the `fname` function
and end up messing the ABI. Therefore, make double sure this is avoided
by restoring the callee preserved registers before switching back to the
initial stack and returning.

Notice that we make use of `TPIDRRO_EL0` as a scratch register so that
we can temporarily preserve x0's value. This should be safe as the
application/user themselves cannot change this register's value,
which means this register's value will always be known by us and we
could restore it anytime.

Some example OSes making use of TPIDRRO_EL0 would include Zephyr using
use this register for holding its `struct __cpu`, Windows using it to
hold the current CPU number, or Linux which uses it as both a scratch
register and a secondary TLS pointer.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>